# Conflicts:
#	src/main/kotlin/it/pagopa/ecommerce/helpdesk/services/PmService.kt
#	src/main/kotlin/it/pagopa/ecommerce/helpdesk/utils/ResponseBuilder.kt
#	src/test/kotlin/it/pagopa/ecommerce/helpdesk/HelpdeskTestUtils.kt
#	src/test/kotlin/it/pagopa/ecommerce/helpdesk/dataproviders/oracle/PMTransactionDataProviderTest.kt

+++ src/main/kotlin/it/pagopa/ecommerce/helpdesk/services/PmService.kt
@@ -36,7 +36,7 @@ class PmService(@Autowired val pmTransactionDataProvider: PMTransactionDataProvi
                     .map { results ->
                         buildTransactionSearchResponse(
                             currentPage = pageNumber,
-                            totalCount = totalCount.toInt(),
+                            totalCount = totalCount,
                             results = results
                         )
                     }
